Added open function to open file and read content to text box
Defined "open" function that will call filedialog for the user to select a text file, and then read that text file into the text box.  It will also save the file location to the global file_location variable.  The open function is connected to the "Open" label in the file menu.